books/requirements.txt
asgiref==3.6.0
coverage==7.2.1
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
